Clarified that all operations in Bits are member functions of the
type class.  Reverse the meaning of the sign of the second argument
for `rotate' and `shift' (this makes it the same as GHC used all
the time).  `bitSize' on `Integer' etc is now undefined.

Allowing multiple calls to hs_init() and clarified the constraints
on the relative timing between hs_set_argv() and
getProgName/getArgs.

darcs-hash:20020911073130-8368b-aaacdc841575729d38fbbe5f6050d8434b77f124